Fix 6 backend bugs: schema nullability, missing selectinload, type mismatches
- models/institution.py: media_gallery type annotation dict→list to match actual usage
- schemas/matching.py: make match_score, match_tier, signal_value nullable to match ORM
- services/application_service.py: add missing selectinload for institution app queries
- schemas/institution.py: add institution_name field to PostResponse (was silently dropped)
- api/institutions.py: remove stale type:ignore now that field exists
- services/student_service.py: wrap Decimal avg_score in float() for correct JSON serialization

+++ unipaith-backend/src/unipaith/services/student_service.py
@@ -922,7 +922,9 @@ class StudentService:
         matches = list(matches_result.scalars().all())
         match_count = len(matches)
         avg_score = (
-            round(sum(m.match_score for m in matches if m.match_score is not None) / match_count, 1)
+            float(round(
+                sum(m.match_score for m in matches if m.match_score is not None) / match_count, 1,
+            ))
             if match_count > 0
             else None
         )
